backend/requirements.txt
django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
pymongo==4.8.0
python-dotenv==1.0.1
qrcode[pil]==7.4.2
user-agents==2.2.0
